fix: Don't drop message for small cache size
fix #2756

Increased route back cache to avoid dropping messages which
are supposed to be routed back. The problem was that when some
nodes are syncing several chunk request are asked, and request
arrive faster than responses, effectively invalidating the cache
most of the  responses are dropped, since the route back hash was
dropped from the cache.

Note: This is a vector of attack if we rely on route-back-messages,
since malicious actor can feed the cache with new information,
effectively invalidating our current cache.

Test plan
=========
Reproduced this scenario locally, and tested with and without the cache.